// src/packages/types/master/Mst_OrderComplainType.ts
export interface Mst_OrderComplainType {
  OrderComplainImageType: string;
  OrderComplainImageName: string;
  FlagActive: string;
}

export interface SearchMst_OrderComplainTypeParam {
  OrderComplainTypeCode: string;
  OrderComplainTypeName: string;
  FlagActive: string;
  Ft_PageIndex: number;
  Ft_PageSize: number;
}